# Several improvements to SocketsHttpHandler perf