@stephentoub 2019-10-08
Which style changes?  Just the `var` to the actual type? I only fixed things where I was already touching the code or where writing new code the "right" way would have left things inconsistent.