@geoffkizer 2019-10-08
Re retrieving header values: It would be nice to avoid the string[] allocation here entirely. This is particularly relevant for HTTP3 since there's no easy way to reuse the string[] across requests -- each request is on its own QUIC stream.

But I don't see an obvious way to do this. Maybe a callback model? We can defer that for now.